enderman stealing fixes

+++ bot/cogs/core/mobs.py
@@ -403,18 +403,22 @@ class MobSpawner(commands.Cog):
                     mob_key == "enderman"
                     and random.randint(1, {"easy": 10, "hard": 5}[difficulty]) == 1
                 ):
-                    stolen_item = random.choice(
-                        [item for item in await self.db.fetch_items(user.id) if item.amount > 1]
-                    )
+                    stealable_items = [item for item in await self.db.fetch_items(user.id) if item.amount > 1 and emojify_item(self.d, item.name) != self.d.emojis.air]
 
-                    await self.db.remove_item(user.id, stolen_item.name, 1)
+                    if len(stealable_items) > 2:
+                        stolen_item = random.choice(stealable_items)
 
-                    await ctx.send_embed(
-                        random.choice(ctx.l.mobs_mech.lost.enderman).format(
-                            1, emojify_item(self.d, stolen_item.name)
+                        await self.db.remove_item(user.id, stolen_item.name, 1)
+
+                        await ctx.send_embed(
+                            random.choice(ctx.l.mobs_mech.lost.enderman).format(
+                                1, emojify_item(self.d, stolen_item.name)
+                            )
                         )
-                    )
-                elif mob_key in {"creeper", "enderman"}:
+
+                        return
+                
+                if mob_key in {"creeper", "enderman"}:
                     await ctx.send_embed(
                         random.choice(ctx.l.mobs_mech.lost[mob_key]).format(
                             ems_lost, self.d.emojis.emerald
